Move GitActions and FileActions into a new Actions module
