+++ lib/homesick.rb
@@ -1,7 +1,7 @@
 # -*- encoding : utf-8 -*-
 require 'homesick/shell'
-require 'homesick/file_actions'
-require 'homesick/git_actions'
+require 'homesick/actions/file_actions'
+require 'homesick/actions/git_actions'
 require 'homesick/version'
 require 'homesick/utils'
 require 'homesick/cli'
