+++ lib/homesick/cli.rb
@@ -5,8 +5,8 @@ module Homesick
   # Homesick's command line interface
   class CLI < Thor
     include Thor::Actions
-    include Homesick::FileActions
-    include Homesick::GitActions
+    include Homesick::Actions::FileActions
+    include Homesick::Actions::GitActions
     include Homesick::Version
     include Homesick::Utils
